Move rest of plexpy to use new url
