+++ util/getNowPlayingFromPlexpy.js
@@ -8,7 +8,6 @@ module.exports = (guildId) =>
         .then((settings) =>
         {   
             let url = `${settings.protocol}://${settings.host}/${settings.path}/api/v2?apikey=${settings.apikey}&cmd=get_activity`;
-            console.log(url);
             request(url, function (e, r, b) {
                 let j = JSON.parse(b);
                 if (e && r.statusCode !== 200) reject(e);
