+++ util/getUserStatsFromPlexpy.js
@@ -1,4 +1,4 @@
-const getSettings = require('./getSettings');
+const getSettings = require('./getPlexPySettings');
 const request = require('request');
 module.exports = (guildId, userId) => 
 {
@@ -7,15 +7,11 @@ module.exports = (guildId, userId) =>
         getSettings(guildId)
         .then((settings) =>
             {
-                settings = JSON.parse(settings);
-                let host = settings.find(x => x.setting == "plexpy.host").value;
-                let baseurl = settings.find(x => x.setting == "plexpy.baseurl").value;
-                let apikey = settings.find(x => x.setting == "plexpy.apikey").value;
-                let url = `http://${host}${baseurl}/api/v2?apikey=${apikey}&cmd=get_user_watch_time_stats&user_id=${userId}`;
+                let url = `${settings.protocol}://${settings.host}/${settings.path}/api/v2?apikey=${settings.apikey}&cmd=get_user_watch_time_stats&user_id=${userId}`;
                 request(url, function (e, r, b) {
                     let j = JSON.parse(b);
                     if (e && r.statusCode !== 200) reject(e);
-                    resolve(JSON.stringify(j.response));
+                    resolve(j.response);
                 });
             }
         );
